Moved track class to it's own file
